requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0
openai>=0.28.0
python-dotenv>=1.0.0
